LPS-84094 Move reusable logic to gradle-util
